// api/src/payment/dtos/process-payment.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
import { IsEmail, IsNotEmpty, IsOptional, IsString } from 'class-validator';

export class ProcessPaymentDto {
  @ApiProperty({
    description: 'Payment token generated by the frontend PSP SDK',
    example: 'tok_visa_success',
  })
  @IsNotEmpty()
  @IsString()
  token: string;

  @ApiProperty({
    description: 'Customer email address',
    example: 'cliente@ejemplo.com',
  })
  @IsNotEmpty()
  @IsEmail()
  customerEmail: string;
}